% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agregat}
\alias{agregat}
\title{Mapowanie kodów teryt i nazw jednostek samorządu terytorialnego}
\format{
\subsection{\code{agregat}}{

Ramka danych (\code{data.frame}) zawierająca 379 wierszy i 5 kolumn:
\describe{
\item{teryt_woj}{Dwucyfrowy numer teryt województwa}
\item{woj_nazwa}{Nazwa województwa}
\item{nazwa_pow_szk}{Nazwa powiatu}
\item{teryt_pow}{Sześciocyfrowy numer teryt powiatu}
\item{podregion}{Nazwa podregionu NUTS3}
}
}
}
\usage{
agregat
}
\description{
Zbiór zawiera mapowanie kodów teryt i nazw jednostek samorządu terytorialnego
od województw, przez podregiony (tylko nazwy) do powiatów. Dane te zostały
wyciągnięte ze zbioru zagregowanych wskaźników z monitoringu absolwentów stąd
brak jednego powiatu. Dane te służą do generowania losowych danych o
strukturze tabeli pośredniej \code{P4} na potrzeby testowania pakietu jako
pierwszy argument funkcji \link{\code{dummyP4}}.
}
\keyword{datasets}
