% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummyP2}
\alias{dummyP2}
\title{Tworzenie dummy data tabeli pośredniej \code{P2}}
\usage{
dummyP2(osobomies, seed = NULL)
}
\arguments{
\item{osobomies}{tabela pośrednia \code{P3} w wersji symulowanej (np. za
pomocą funkcji \link{\code{dummyP3}})}

\item{seed}{wartość ziarna generatora - domyślna wartość to \code{NULL}
(argument opcjonalny na wypadek potrzeby powtarzalności obliczeń)}
}
\value{
data.frame
}
\description{
Funkcja tworzy dummy data tabeli pośredniej \code{P2} na
podstawie tabeli pośredniej \code{P3} (zbiór osobo-miesięcy opisujących
sytuację edukacyjno-zawodową absolwentów), która potem będzie używana do
testowania pakietu \code{MLASZdaneAdm4}.
}
\details{
Zbiór tworzony przez tę funkcję ma odzwierciedlać tabelę pośrednią
\code{P2} w stopniu, jaki jest wymagany do przetestowania funkcji liczących
wskaźniki zagregowane. Tabela \code{P3}, a konkretniej jej symulowana wersja
tworzona przez funkcję \link{\code{dummyP3}}, to tabela osobo-miesięcy
opisujących sytuację edukacyjno-zawodową absolwentów w poszczególnych
miesiącach.
Zbiór tworzony przez funkcję zawiera 3 testowane w pakiecie
\code{MLASZdaneAdm4} zmienne:
\itemize{
\item{branza_kont}{dla absolwentów kontynuujących naukę w szkołach
kształcących w zawodzie jest to branża, do której należy zawód, którego uczą
się w tej szkole}
\item{dziedzina_kont}{dla absolwentów kontynuujących kształcenie na studiach
jest to dziedzina naukowa, do której należy kierunek, który studiują}
\item{dyscyplina_wiodaca_kont}{odpowiadająca dziedzinie dyscyplina naukowa}
}
}
