% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummyP4}
\alias{dummyP4}
\title{Tworzenie dummy data na podstawie tabeli pośredniej \code{P4}}
\usage{
dummyP4(agregat, rspo, seed = NULL)
}
\arguments{
\item{agregat}{agregat, na podstawie, którego tworzony będzie zbiór (w
rzeczywistości jest to mapowanie numerów teryt i nazw JST na podstawie tabeli
pośredniej \code{P4})}

\item{rspo}{zbiór RSPO zawierający minimum 100 wierszy opisujących id RSPO,
nazwę i adres szkoły}

\item{seed}{wartość ziarna generatora - domyślna wartość to \code{NULL}
(argument opcjonalny na wypadek potrzeby powtarzalności obliczeń)}
}
\value{
data.frame
}
\description{
Funkcja tworzy dummy data na podstawie tabeli pośredniej
\code{P4}, które potem będą używane do testowania pakietu
\code{MLASZdaneAdm4}.
}
\details{
Zbiór tworzony przez tę funkcję ma odzwierciedlać tabelę pośrednią
\code{P4} w stopniu, jaki jest wymagany do przetestowania funkcji liczących
wskaźniki zagregowane. Ze względu na to, że dane z monitoringu nie mogą być w
formie indywidualnej dostępne publicznie, a na potrzeby testowania pakietu
potrzebny jest jakiś zbiór odzwierciedlający strukturę prawdziwego zbioru, na
potrzeby testowania pakietu tworzony będzie zbiór dummy data. Zbiór ten
będzie zawierał zmienne podobne do tych w zbiorze z monitoringu, ale nie
zawsze będzie zachowana logiczna zależność między zmiennymi. Na przykład,
nazwy szkoły mogą nie odpowiadać typowi szkoły. Jest to celowe uproszczenie,
ponieważ ten aspekt nie będzie podlegał testowaniu, ale wymienione zmienne są
wymagane na potrzeby testów. Takich uproszczeń w zbiorze jest więcej, ale
zostały zastosowane tylko tam, gdzie ma to sens.
Uproszczenia te wynikają z potrzeby generowania zbioru za każdym razem kiedy
testy są uruchamiane. Dzięki temu, że przy każdym testowaniu używany jest
nieco inny zbiór ograniczany jest wpływ zawartości zbioru na testy. Przy raz
wygenerowanym zbiorze jakiś specyficzny układ wartości zmiennych mógłby mieć
wpływ na test, a przy generowaniu wartości zmiennych oddzielnie dla każdego
testu wpływ ten jest ograniczany.
}
