% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rspo}
\alias{rspo}
\title{Próbka nazw i adresów szkół z rejestru RSPO}
\format{
\subsection{\code{rspo}}{

Ramka danych (\code{data.frame}) zawierająca 100 wierszy i 3 kolumny:
\describe{
\item{id_szk}{Numer RSPO szkoły}
\item{nazwa_szk}{Nazwa szkoły z rejestru RSPO}
\item{adres_szk}{Adres szkoły z rejestru RSPO}
}
}
}
\source{
\url{https://rspo.gov.pl/}
}
\usage{
rspo
}
\description{
Wycinek ogólnodostępnego zbioru Rejestru Szkół i Placówek Oświatowych (RSPO).
Zbiór danych zawiera wyslosowane 100 szkół. Dane te służą do generowania
losowych danych o strukturze tabeli pośredniej \code{P4} na potrzeby testowania
pakietu jako drugi argument funkcji \link{\code{dummyP4}}.
}
\keyword{datasets}
