% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_wskazniki_adm4.R
\name{dane_szkoly}
\alias{dane_szkoly}
\title{Wskaźniki zagregowane dla 4. edycji monitoringu - dane administracyjne}
\usage{
dane_szkoly(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
list
}
\description{
Funkcja przechowująca nazwę i adres (z RSPO) szkoły, której
dotyczy raport.
}
